Add function to start system watcher on the end of the script

+++ Powershell/Security/Security_and_Cleanup/Security_and_Cleanup.ps1
@@ -262,9 +262,9 @@ function Set-Hostname() {
 		Write-Output('[i] SetHostname')
 
 		try {
-			$script:controllerId = (Get-ItemProperty -Path 'Registry::HKLM\SOFTWARE\EZUniverse\EZ360ControllerInstaller' -Name 'ControllerID' -ErrorAction SilentlyContinue).ControllerID
+			$script:controllerId = (Invoke-Sqlcmd -ConnectionString $connectionStringEz360 -Query $getControllerIdQuery -QueryTimeout 30 -ErrorAction Stop).ControllerId
 			if ($null -eq $script:controllerId) {
-				$script:controllerId = (Invoke-Sqlcmd -ConnectionString $connectionStringEz360 -Query $getControllerIdQuery -QueryTimeout 30 -ErrorAction Stop).ControllerId
+				$script:controllerId = (Get-ItemProperty -Path 'Registry::HKLM\SOFTWARE\EZUniverse\EZ360ControllerInstaller' -Name 'ControllerID' -ErrorAction SilentlyContinue).ControllerID
 			}
 		} catch {
 			$script:gatheredErrors += ("[-]  $($_.Exception.Message)")
@@ -402,9 +402,29 @@ function Uninstall-LegacyComponents() {
 	Get-ChildItem |
 	Get-ItemProperty
 
-	Write-Output('[i] Stopping SystemWatcher, SQLReplicator, EZSensor Server and UpdateCenters')
-	Stop-Service -Force -ErrorAction SilentlyContinue -Name (
-		'EZSQLReplicator','EZSystemWatcher', 'EZSensorsServer','SubwayUpdateCenter','EZUpdateCenter')
+	$appNames = $appObjects | ForEach-Object {$_.DisplayName}
+
+	$isAnythingToBeRemoved = $False
+	foreach ($app in $allApplications) {
+		if ($appNames.Contains($app)) {
+			$isAnythingToBeRemoved = $True
+			break
+		}
+	}
+
+	if ($isAnythingToBeRemoved -eq $False) {return}
+
+	foreach ($service in $servicesToRemove) {
+		if ($appNames.Contains($service)) {
+			Stop-Service -Force -ErrorAction SilentlyContinue -Name (
+				'EZSQLReplicator','EZSystemWatcher', 'EZSensorsServer','SubwayUpdateCenter','EZUpdateCenter')
+			Write-Output('[i] Stopping SystemWatcher, SQLReplicator, EZSensor Server and UpdateCenters')
+			break
+		}
+	}
+
+	
+	
 
 	if (Get-Service -Name 'EZSensors Server' -ErrorAction SilentlyContinue) {
 		Set-Service -Name 'EZSensors Server' -Force -StartupType Disabled
@@ -480,6 +500,7 @@ function Uninstall-LegacyComponents() {
 			$script:gatheredErrors += ("[-]  $($_.Exception.Message)")
 		}
 	}
+	Ensure-EZSystemWatcherIsRunning
 	Write-Output('[+] Finished removing Legacy components')
 }
 
@@ -770,13 +791,39 @@ function Limit-AccessToFolders() {
 	Add-NTFSAccess -Path 'C:\DTIQ\Security_and_Cleanup' -Account 'Administrators' -AccessRights Full -InheritanceFlags ObjectInherit
 }
 
+function Ensure-EZSystemWatcherIsRunning () {
+	function Check-ControllerInstallerIsRunning () {
+		if (Get-Process -Name '360IQControllerInstaller*' -ErrorAction SilentlyContinue) {
+			return $true
+		}
+		else {
+			return $false
+		}
+	}
+	# Installer After finishing its current run will start EZSystemWatcher
+	if (Check-ControllerInstallerIsRunning) {return}
+
+	$Watcher = Get-Service -Name 'EZSystemWatcher' -ErrorAction SilentlyContinue
+	$retryCount = 5
+
+
+	while (($Watcher.Status -ne 'Running') -and ($retryCount -ne 0)) {
+		Start-Service $Watcher -ErrorAction SilentlyContinue
+		Start-Sleep -Seconds 3
+		$retryCount = $retryCount - 1
+	}
+
+	if (($Watcher | Get-Service).Running -ne 'Running') {
+		$script:gatheredErrors += ("[-]  $($_.Exception.Message)")
+	}
+}
+
 $startTime = Get-Date
 Install-Packages
 Limit-AccessToFolders
 Get-DotNetFiles
 Disable-Windows11Upgrade
 Set-Hostname
-# Install-Atera # ! Moved to separate file
 Disable-WindowsUpdateIfAteraNotPresent
 Disable-Obee
 Set-FirewallRulePingAllow
@@ -785,7 +832,7 @@ Install-Wazuh
 Install-DotNet
 Disable-Copilot
 Move-DotnetEnvVarPosition
-Disable-Copilot
+Ensure-EZSystemWatcherIsRunning
 Push-ErrorLogs($script:gatheredErrors)
 
 <# TODO :
